reorg executes
